+++ RELEASE_NOTES.md
@@ -1,3 +1,13 @@
+#### 1.0.3 - 2018-01-09
+* added: 
+    * Stream.WriteJob
+    * Stream.CopyToJob
+    * HttpWebRequest.GetRequestStreamJob
+    * Azure.getAdTokenJob
+    * Datalake.uploadStreamJob
+    * Datalake.downloadStreamJob
+    * Datalake.downloadContentJob
+
 #### 1.0.2 - 2017-12-21
 * HttpWebRequest added